dunno... some refactoring?

+++ app/src/main/java/com/daveloper/curriculumvitae/data/entities/Language.kt
@@ -6,5 +6,5 @@ import androidx.room.PrimaryKey
 @Entity
 data class Language(
     @PrimaryKey val name: String,
-    val classification: Int
+    val classification: Float
 )
